+++ sdk/src/main/kotlin/dev/datastar/kotlin/sdk/ServerSentEventGenerator.kt
@@ -5,9 +5,17 @@ package dev.datastar.kotlin.sdk
  * Common interface for the Datastar SDK with which adapters must comply.
  */
 interface Request {
+    enum class Method {
+        GET,
+        POST,
+        PUT,
+        PATCH,
+        DELETE,
+    }
+
     fun bodyString(): String
 
-    fun isGet(): Boolean
+    fun method(): Method
 
     fun readParam(string: String): String
 }
@@ -166,11 +174,13 @@ data class SendEventOptions(
 
 const val DEFAULT_RETRY_DURATION = 1000L
 val DEFAULT_MODE = ElementPatchMode.Outer
+val DEFAULT_NAMESPACE = ElementNamespace.Html
 
 data class PatchElementsOptions(
     val selector: String? = null,
     val mode: ElementPatchMode = DEFAULT_MODE,
     val useViewTransition: Boolean = false,
+    val namespace: ElementNamespace = DEFAULT_NAMESPACE,
     val eventId: String? = null,
     val retryDuration: Long = DEFAULT_RETRY_DURATION,
 )
@@ -191,6 +201,19 @@ enum class ElementPatchMode(
 
     companion object {
         operator fun invoke(value: String?) = entries.firstOrNull { it.value == value } ?: DEFAULT_MODE
+    }
+}
+
+enum class ElementNamespace(
+    val value: String,
+) {
+    Html("html"),
+    Svg("svg"),
+    MathMl("mathml"),
+    ;
+
+    companion object {
+        operator fun invoke(value: String?) = entries.firstOrNull { it.value == value } ?: DEFAULT_NAMESPACE
     }
 }
 
@@ -247,6 +270,7 @@ private class ServerSentEventGeneratorBase(
                 options.selector?.let { add("selector $it") }
                 options.mode.let { if (it != DEFAULT_MODE) add("mode ${it.value}") }
                 options.useViewTransition.let { if (it) add("useViewTransition ${true}") }
+                options.namespace.let { if (it != DEFAULT_NAMESPACE) add("namespace ${it.value}") }
                 (elements ?: "")
                     .lineSequence()
                     .filter { it.isNotBlank() }